refactor: 이벤트 좋아요 엔드포인트를 EventController로 이동
/events/{eventId}/like 관련 엔드포인트들을 별도의 LikeController에서
EventController로 이동하였습니다. 이를 통해 동일한 리소스에 대한
모든 작업이 하나의 컨트롤러에서 관리되도록 구조를 개선하였습니다.

- 기존 LikeController에서 구현된 likeEvent, unlikeEvent 메서드를
 EventController로 이동
- URL 경로와 컨트롤러 구조 간 일관성 향상
- 리소스 중심의 컨트롤러 구조 확립
