+++ src/main/java/com/moonbaar/domain/event/controller/EventController.java
@@ -5,11 +5,16 @@ import com.moonbaar.domain.event.dto.EventListResponse;
 import com.moonbaar.domain.event.dto.EventSearchRequest;
 import com.moonbaar.domain.event.exeption.InvalidEventParamsException;
 import com.moonbaar.domain.event.service.EventService;
+import com.moonbaar.domain.like.dto.LikeResponse;
+import com.moonbaar.domain.like.service.LikeService;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,7 +24,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class EventController {
 
+    // 임시 유저 아이디
+    private Long MOCK_USER_ID = 1L;
     private final EventService eventService;
+    private final LikeService likeService;
 
     @GetMapping
     public EventListResponse searchEvents(
@@ -79,5 +87,17 @@ public class EventController {
         if (page < 1 || size < 1 || size > 100) {
             throw new InvalidEventParamsException();
         }
+    }
+
+    @PostMapping("/{eventId}/like")
+    public ResponseEntity<LikeResponse> likeEvent(@PathVariable Long eventId) {
+        LikeResponse response = likeService.likeEvent(MOCK_USER_ID, eventId);
+        return ResponseEntity.ok(response);
+    }
+
+    @DeleteMapping("/{eventId}/like")
+    public ResponseEntity<LikeResponse> unlikeEvent(@PathVariable Long eventId) {
+        LikeResponse response = likeService.unlikeEvent(MOCK_USER_ID, eventId);
+        return ResponseEntity.ok(response);
     }
 }
